# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Day Trading Tool API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Quote {
  symbol: string;
  name: string;
  price: number;
  change: number;
  changePercent: number;
  volume: number;
  /** @nullable */
  marketCap: number | null;
  open: number;
  high: number;
  low: number;
  previousClose: number;
  fiftyTwoWeekHigh: number;
  fiftyTwoWeekLow: number;
  /** @nullable */
  avgVolume?: number | null;
  /** @nullable */
  pe?: number | null;
  /** @nullable */
  eps?: number | null;
  updatedAt: string;
}

export interface Candle {
  timestamp: number;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}

export interface HistoryResponse {
  symbol: string;
  interval?: string;
  range?: string;
  candles: Candle[];
}

export interface OptionContract {
  contractSymbol: string;
  strike: number;
  expiry: string;
  lastPrice: number;
  bid: number;
  ask: number;
  impliedVolatility: number;
  openInterest: number;
  volume: number;
  /** @nullable */
  delta: number | null;
  /** @nullable */
  gamma: number | null;
  /** @nullable */
  theta: number | null;
  /** @nullable */
  vega: number | null;
  inTheMoney: boolean;
}

export interface OptionExpiry {
  date: string;
  calls: OptionContract[];
  puts: OptionContract[];
}

export interface OptionChain {
  symbol: string;
  underlyingPrice: number;
  expiryDates: string[];
  options: OptionExpiry[];
}

export interface SearchResult {
  symbol: string;
  name: string;
  type: string;
  exchange: string;
}

export interface MoverItem {
  symbol: string;
  name: string;
  price: number;
  change: number;
  changePercent: number;
  volume: number;
}

export interface MarketMovers {
  gainers: MoverItem[];
  losers: MoverItem[];
  active: MoverItem[];
}

export interface WatchlistItem {
  id: number;
  symbol: string;
  name: string;
  addedAt: string;
}

export interface WatchlistInput {
  symbol: string;
  name: string;
}

export type PositionAssetType = typeof PositionAssetType[keyof typeof PositionAssetType];


export const PositionAssetType = {
  stock: 'stock',
  option: 'option',
} as const;

export interface Position {
  id: number;
  symbol: string;
  name: string;
  quantity: number;
  avgCost: number;
  openedAt: string;
  /** @nullable */
  notes?: string | null;
  assetType: PositionAssetType;
  /** @nullable */
  contractType?: string | null;
  /** @nullable */
  strikePrice?: number | null;
  /** @nullable */
  expiryDate?: string | null;
}

export type PositionInputAssetType = typeof PositionInputAssetType[keyof typeof PositionInputAssetType];


export const PositionInputAssetType = {
  stock: 'stock',
  option: 'option',
} as const;

export interface PositionInput {
  symbol: string;
  name: string;
  quantity: number;
  avgCost: number;
  notes?: string;
  assetType: PositionInputAssetType;
  contractType?: string;
  strikePrice?: number;
  expiryDate?: string;
}

export interface PositionUpdate {
  quantity?: number;
  avgCost?: number;
  notes?: string;
}

export interface PortfolioSummary {
  totalCost: number;
  totalValue: number;
  totalPnl: number;
  totalPnlPercent: number;
  dayPnl: number;
  dayPnlPercent: number;
  positionCount: number;
}

export type TradeSide = typeof TradeSide[keyof typeof TradeSide];


export const TradeSide = {
  buy: 'buy',
  sell: 'sell',
} as const;

export type TradeAssetType = typeof TradeAssetType[keyof typeof TradeAssetType];


export const TradeAssetType = {
  stock: 'stock',
  option: 'option',
} as const;

export interface Trade {
  id: number;
  symbol: string;
  side: TradeSide;
  quantity: number;
  price: number;
  fees?: number;
  /** @nullable */
  pnl?: number | null;
  /** @nullable */
  notes?: string | null;
  tradedAt: string;
  assetType: TradeAssetType;
  /** @nullable */
  contractType?: string | null;
  /** @nullable */
  strikePrice?: number | null;
  /** @nullable */
  expiryDate?: string | null;
}

export type TradeInputSide = typeof TradeInputSide[keyof typeof TradeInputSide];


export const TradeInputSide = {
  buy: 'buy',
  sell: 'sell',
} as const;

export type TradeInputAssetType = typeof TradeInputAssetType[keyof typeof TradeInputAssetType];


export const TradeInputAssetType = {
  stock: 'stock',
  option: 'option',
} as const;

export interface TradeInput {
  symbol: string;
  side: TradeInputSide;
  quantity: number;
  price: number;
  fees?: number;
  pnl?: number;
  notes?: string;
  tradedAt: string;
  assetType: TradeInputAssetType;
  contractType?: string;
  strikePrice?: number;
  expiryDate?: string;
}

export interface TradeUpdate {
  price?: number;
  quantity?: number;
  fees?: number;
  pnl?: number;
  notes?: string;
  tradedAt?: string;
}

export interface TradeStats {
  totalTrades: number;
  wins: number;
  losses: number;
  winRate: number;
  avgWin: number;
  avgLoss: number;
  totalPnl: number;
  profitFactor: number;
  largestWin: number;
  largestLoss: number;
  avgHoldingPeriodHours: number;
}

export type AlertCondition = typeof AlertCondition[keyof typeof AlertCondition];


export const AlertCondition = {
  above: 'above',
  below: 'below',
} as const;

export interface Alert {
  id: number;
  symbol: string;
  condition: AlertCondition;
  targetPrice: number;
  /** @nullable */
  note?: string | null;
  isActive: boolean;
  createdAt: string;
  /** @nullable */
  triggeredAt?: string | null;
}

export type AlertInputCondition = typeof AlertInputCondition[keyof typeof AlertInputCondition];


export const AlertInputCondition = {
  above: 'above',
  below: 'below',
} as const;

export interface AlertInput {
  symbol: string;
  condition: AlertInputCondition;
  targetPrice: number;
  note?: string;
}

export type AlertUpdateCondition = typeof AlertUpdateCondition[keyof typeof AlertUpdateCondition];


export const AlertUpdateCondition = {
  above: 'above',
  below: 'below',
} as const;

export interface AlertUpdate {
  targetPrice?: number;
  condition?: AlertUpdateCondition;
  note?: string;
  isActive?: boolean;
}

export interface ScanResult {
  symbol: string;
  name: string;
  price: number;
  change: number;
  changePercent: number;
  volume: number;
  /** @nullable */
  marketCap: number | null;
  /** @nullable */
  pe?: number | null;
}

export interface DashboardSummary {
  portfolioValue: number;
  portfolioPnl: number;
  portfolioPnlPercent: number;
  dayPnl: number;
  dayPnlPercent: number;
  activeAlerts: number;
  triggeredAlerts: number;
  totalTrades: number;
  recentTrades: Trade[];
  topMovers: MoverItem[];
}

export type GetHistoryParams = {
symbol: string;
interval?: GetHistoryInterval;
range?: GetHistoryRange;
};

export type GetHistoryInterval = typeof GetHistoryInterval[keyof typeof GetHistoryInterval];


export const GetHistoryInterval = {
  '1m': '1m',
  '5m': '5m',
  '15m': '15m',
  '30m': '30m',
  '1h': '1h',
  '1d': '1d',
  '1wk': '1wk',
} as const;

export type GetHistoryRange = typeof GetHistoryRange[keyof typeof GetHistoryRange];


export const GetHistoryRange = {
  '1d': '1d',
  '5d': '5d',
  '1mo': '1mo',
  '3mo': '3mo',
  '6mo': '6mo',
  '1y': '1y',
  '2y': '2y',
  '5y': '5y',
} as const;

export type GetOptionChainParams = {
symbol: string;
expiry?: string;
};

export type SearchSymbolsParams = {
q: string;
};

export type GetTradesParams = {
symbol?: string;
limit?: number;
offset?: number;
};

export type ScanStocksParams = {
minPrice?: number;
maxPrice?: number;
minVolume?: number;
minChangePercent?: number;
maxChangePercent?: number;
/**
 * Preset scan: gainers, losers, high-volume, gap-up, gap-down
 */
preset?: string;
};