# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Day Trading Tool API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /market/quote/{symbol}:
    get:
      operationId: getQuote
      tags: [market]
      summary: Get real-time quote for a symbol
      parameters:
        - name: symbol
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Quote data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "404":
          description: Symbol not found
  /market/history:
    get:
      operationId: getHistory
      tags: [market]
      summary: Get OHLCV price history
      parameters:
        - name: symbol
          in: query
          required: true
          schema: { type: string }
        - name: interval
          in: query
          required: false
          schema:
            type: string
            enum: [1m, 5m, 15m, 30m, 1h, 1d, 1wk]
            default: 1d
        - name: range
          in: query
          required: false
          schema:
            type: string
            enum: [1d, 5d, 1mo, 3mo, 6mo, 1y, 2y, 5y]
            default: 3mo
      responses:
        "200":
          description: Historical OHLCV data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryResponse"
  /market/options:
    get:
      operationId: getOptionChain
      tags: [market]
      summary: Get options chain for a symbol
      parameters:
        - name: symbol
          in: query
          required: true
          schema: { type: string }
        - name: expiry
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Option chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OptionChain"
  /market/search:
    get:
      operationId: searchSymbols
      tags: [market]
      summary: Search for ticker symbols
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchResult"
...
  /portfolio:
    get:
      operationId: getPortfolioPositions
      tags: [portfolio]
      summary: Get all portfolio positions
      responses:
        "200":
          description: Positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Position"
    post:
      operationId: addPosition
      tags: [portfolio]
      summary: Add a new position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PositionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
  /portfolio/{id}:
    patch:
      operationId: updatePosition
      tags: [portfolio]
      summary: Update a position
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PositionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
    delete:
      operationId: deletePosition
      tags: [portfolio]
      summary: Delete a position
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /alerts:
    get:
      operationId: getAlerts
      tags: [alerts]
      summary: Get all price alerts
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
    post:
      operationId: createAlert
      tags: [alerts]
      summary: Create a price alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}:
    patch:
      operationId: updateAlert
      tags: [alerts]
      summary: Update an alert
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
    delete:
      operationId: deleteAlert
      tags: [alerts]
      summary: Delete an alert
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Scanner ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Quote:
      type: object
      required: [symbol, name, price, change, changePercent, volume, marketCap, open, high, low, previousClose, fiftyTwoWeekHigh, fiftyTwoWeekLow, updatedAt]
      properties:
        symbol:         { type: string }
        name:           { type: string }
        price:          { type: number }
        change:         { type: number }
        changePercent:  { type: number }
        volume:         { type: integer }
        marketCap:      { type: ["number", "null"] }
        open:           { type: number }
        high:           { type: number }
        low:            { type: number }
        previousClose:  { type: number }
        fiftyTwoWeekHigh: { type: number }
        fiftyTwoWeekLow:  { type: number }
        avgVolume:      { type: ["integer", "null"] }
        pe:             { type: ["number", "null"] }
        eps:            { type: ["number", "null"] }
        updatedAt:      { type: string, format: date-time }
    Candle:
      type: object
      required: [timestamp, open, high, low, close, volume]
      properties:
        timestamp: { type: integer }
        open:      { type: number }
        high:      { type: number }
        low:       { type: number }
        close:     { type: number }
        volume:    { type: integer }
    HistoryResponse:
      type: object
      required: [symbol, candles]
      properties:
        symbol:   { type: string }
        interval: { type: string }
        range:    { type: string }
        candles:
          type: array
          items:
            $ref: "#/components/schemas/Candle"
    OptionContract:
      type: object
      required: [contractSymbol, strike, expiry, lastPrice, bid, ask, impliedVolatility, openInterest, volume, delta, gamma, theta, vega, inTheMoney]
      properties:
        contractSymbol:     { type: string }
        strike:             { type: number }
        expiry:             { type: string }
        lastPrice:          { type: number }
        bid:                { type: number }
        ask:                { type: number }
        impliedVolatility:  { type: number }
        openInterest:       { type: integer }
        volume:             { type: integer }
        delta:              { type: ["number", "null"] }
        gamma:              { type: ["number", "null"] }
        theta:              { type: ["number", "null"] }
        vega:               { type: ["number", "null"] }
        inTheMoney:         { type: boolean }
    OptionExpiry:
      type: object
      required: [date, calls, puts]
      properties:
        date:  { type: string }
        calls:
          type: array
          items:
            $ref: "#/components/schemas/OptionContract"
        puts:
          type: array
          items:
            $ref: "#/components/schemas/OptionContract"
    OptionChain:
      type: object
      required: [symbol, underlyingPrice, expiryDates, options]
      properties:
        symbol:          { type: string }
        underlyingPrice: { type: number }
        expiryDates:
          type: array
          items: { type: string }
        options:
          type: array
          items:
            $ref: "#/components/schemas/OptionExpiry"
    SearchResult:
      type: object
      required: [symbol, name, type, exchange]
      properties:
        symbol:   { type: string }
        name:     { type: string }
        type:     { type: string }
        exchange: { type: string }
...
    Position:
      type: object
      required: [id, symbol, name, quantity, avgCost, openedAt, assetType]
      properties:
        id:        { type: integer }
        symbol:    { type: string }
        name:      { type: string }
        quantity:  { type: number }
        avgCost:   { type: number }
        openedAt:  { type: string, format: date-time }
        notes:     { type: ["string", "null"] }
        assetType:
          type: string
          enum: [stock, option]
        contractType: { type: ["string", "null"] }
        strikePrice:  { type: ["number", "null"] }
        expiryDate:   { type: ["string", "null"] }
    PositionInput:
      type: object
      required: [symbol, name, quantity, avgCost, assetType]
      properties:
        symbol:    { type: string }
        name:      { type: string }
        quantity:  { type: number }
        avgCost:   { type: number }
        notes:     { type: string }
        assetType:
          type: string
          enum: [stock, option]
        contractType: { type: string }
        strikePrice:  { type: number }
        expiryDate:   { type: string }
    PositionUpdate:
      type: object
      properties:
        quantity:  { type: number }
        avgCost:   { type: number }
        notes:     { type: string }
...
    Alert:
      type: object
      required: [id, symbol, condition, targetPrice, isActive, createdAt]
      properties:
        id:           { type: integer }
        symbol:       { type: string }
        condition:
          type: string
          enum: [above, below]
        targetPrice:  { type: number }
        note:         { type: ["string", "null"] }
        isActive:     { type: boolean }
        createdAt:    { type: string, format: date-time }
        triggeredAt:  { type: ["string", "null"] }
    AlertInput:
      type: object
      required: [symbol, condition, targetPrice]
      properties:
        symbol:      { type: string }
        condition:
          type: string
          enum: [above, below]
        targetPrice: { type: number }
        note:        { type: string }
    AlertUpdate:
      type: object
      properties:
        targetPrice: { type: number }
        condition:
          type: string
          enum: [above, below]
        note:        { type: string }
        isActive:    { type: boolean }

    # ── Scanner ───────────────────────────────────────────────────────────────
...